requirements.txt
asyncpg~=0.29.0
dearpygui==1.10.0
fastapi~=0.104.1
fastapi-utils==0.2.1
httpx==0.25.1
matplotlib~=3.8.0
python-dotenv~=1.0.0
uvicorn~=0.24.0.post1
